move component state logic to hook file
