+++ ui/src/components/ServicePage.jsx
@@ -1,69 +1,16 @@
-import axios from "axios";
-import { useCallback, useEffect, useState } from "react";
 import { Link, useParams } from "react-router-dom";
-import { useDataContext } from "../contexts/DataContext";
 
 import "../styles/components/ServicePage.css";
 import { useNavbarContext } from "../contexts/NavbarContext";
-import { API_GURU_BASE_URL } from "../utils/constants";
+import { useFetchServiceData } from "../utils/useFetchServiceData";
 
 export const ServicePage = () => {
   const { provider, service } = useParams();
-  const { data, loading } = useDataContext();
   const { setNavbarToggle } = useNavbarContext();
-
-  const [serviceData, setServiceData] = useState(null);
-  const [infoList, setInfoList] = useState(null);
-  const [noDataFound, setNoDataFound] = useState(false);
-
-  const iterateThroughObject = (obj) => {
-    const keyValuePairs = [];
-    Object.entries(obj).forEach(([key, value]) => {
-      keyValuePairs.push({
-        key: key,
-        value: value,
-      });
-    });
-    return keyValuePairs;
-  };
-
-  const setPageData = useCallback((service, apiMap) => {
-    if (service in apiMap) {
-      const serviceData = apiMap[service];
-      setServiceData(serviceData);
-      if ("contact" in serviceData.info) {
-        const infoList = iterateThroughObject(serviceData.info.contact);
-        setInfoList(infoList);
-      }
-    } else {
-      setNoDataFound(true);
-    }
-  }, []);
-
-  useEffect(() => {
-    if (!loading && data) {
-      let providerData = null;
-      data.forEach((datum) => {
-        if (datum.name === provider) {
-          providerData = datum;
-          return;
-        }
-      });
-      if (providerData) {
-        setPageData(service, providerData.apis);
-      } else {
-        axios
-          .get(`${API_GURU_BASE_URL}/${provider}.json`)
-          .then((res) => {
-            setPageData(service, res.data.apis);
-          })
-          .catch((error) => {
-            setNoDataFound(true);
-            console.error(error);
-          });
-      }
-    }
-  }, [data, loading, provider, service, setPageData]);
+  const { serviceData, infoList, noDataFound } = useFetchServiceData(
+    provider,
+    service,
+  );
 
   return (
     <>
